+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityArrow.java
@@ -1,7 +1,8 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityArrowBaseComponent;
+import org.allaymc.api.entity.component.EntityDamageComponent;
+import org.allaymc.api.entity.component.attribute.EntityAttributeComponent;
 
-public interface EntityArrow extends Entity {
-
+public interface EntityArrow extends EntityProjectile, EntityArrowBaseComponent, EntityDamageComponent, EntityAttributeComponent {
 }