+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityProjectilePhysicsComponentImpl.java
@@ -6,6 +6,7 @@ import org.allaymc.api.entity.component.EntityAgeComponent;
 import org.allaymc.api.entity.component.EntityProjectileComponent;
 import org.allaymc.api.entity.interfaces.EntityProjectile;
 import org.allaymc.api.eventbus.event.entity.ProjectileHitEvent;
+import org.allaymc.api.math.MathUtils;
 import org.allaymc.api.math.location.Location3d;
 import org.allaymc.api.math.position.Position3i;
 import org.allaymc.server.component.annotation.ComponentObject;
@@ -119,6 +120,11 @@ public class EntityProjectilePhysicsComponentImpl extends EntityPhysicsComponent
             newPos.add(motion.mul(rayCastResult.result, new Vector3d()));
         }
 
+        if (newPos.distance(location) > 0) {
+            // Compute rotation based on the motion only when the projectile moved
+            computeRotationFromMotion(newPos, this.motion);
+        }
+
         if (!newPos.equals(location) && thisEntity.trySetLocation(newPos)) {
             if (rayCastResult.hit instanceof Block block && callHitEvent(newPos, null, block)) {
                 block.getBehavior().onProjectileHit(block, (EntityProjectile) thisEntity, newPos);
@@ -132,6 +138,11 @@ public class EntityProjectilePhysicsComponentImpl extends EntityPhysicsComponent
         }
 
         return false;
+    }
+
+    protected void computeRotationFromMotion(Location3d pos, Vector3dc motion) {
+        pos.setYaw(-MathUtils.getYawFromVector(motion));
+        pos.setPitch(-MathUtils.getPitchFromVector(motion));
     }
 
     protected boolean callHitEvent(Vector3dc hitPos, Entity victim, Block block) {