+++ server/src/main/java/org/allaymc/server/entity/impl/EntityArrowImpl.java
@@ -1,14 +1,36 @@
 package org.allaymc.server.entity.impl;
 
-import java.util.List;
+import lombok.experimental.Delegate;
 import org.allaymc.api.component.interfaces.Component;
+import org.allaymc.api.entity.component.*;
+import org.allaymc.api.entity.component.attribute.EntityAttributeComponent;
 import org.allaymc.api.entity.initinfo.EntityInitInfo;
 import org.allaymc.api.entity.interfaces.EntityArrow;
 import org.allaymc.server.component.interfaces.ComponentProvider;
 
+import java.util.List;
+
 public class EntityArrowImpl extends EntityImpl implements EntityArrow {
+
+    @Delegate
+    protected EntityAgeComponent ageComponent;
+    @Delegate
+    protected EntityProjectileComponent projectileComponent;
+    @Delegate
+    protected EntityPhysicsComponent physicsComponent;
+    @Delegate
+    protected EntityDamageComponent damageComponent;
+    @Delegate
+    protected EntityAttributeComponent attributeComponent;
+
     public EntityArrowImpl(EntityInitInfo initInfo,
-            List<ComponentProvider<? extends Component>> componentProviders) {
-        super(initInfo, componentProviders);;
+                           List<ComponentProvider<? extends Component>> componentProviders) {
+        super(initInfo, componentProviders);
+    }
+
+    @Delegate
+    @Override
+    public EntityArrowBaseComponent getBaseComponent() {
+        return (EntityArrowBaseComponent) super.getBaseComponent();
     }
 }