+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 - (API) Added PlayerInteractBlockEvent.
 - (API) Added PlayerBookEditEvent.
+- (API) Implemented arrow and bow. A new event `EntityShootBowEvent` is added.
+- (API) Introduced methods `Entity.getOffsetAABBForCollisionCheck()` and `Entity.checkBlockCollision()`.
 
 ### Fixed